nft-query/requirements.txt
fastapi
uvicorn
web3
python-dotenv
